+++ zerver/tests/test_audit_log.py
@@ -1156,7 +1156,7 @@ class TestRealmAuditLog(ZulipTestCase):
             [hamlet, cordelia],
             acting_user=hamlet,
             description="lorem",
-            group_settings_map={"can_mention_groups": public_group},
+            group_settings_map={"can_mention_groups": [public_group]},
         )
 
         audit_log_entries = RealmAuditLog.objects.filter(